python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Wycliffe-Kim <f16o@hanyang.ac.kr>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.41.1"
grpcio-tools = "^1.41.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "python.main:main"